Consolidate 9 data model headers into 2 files (models.h, generated_models.h)
