+++ tooling/pipeline/includes/biergarten_data_generator.h
@@ -11,9 +11,7 @@
 #include <vector>
 
 #include "data_generation/data_generator.h"
-#include "data_model/enriched_city.h"
-#include "data_model/generated_brewery.h"
-#include "data_model/location.h"
+#include "data_model/generated_models.h"
 #include "services/enrichment/enrichment_service.h"
 #include "services/database/export_service.h"
