+++ tooling/pipeline/includes/data_generation/llama_generator.h
@@ -16,7 +16,7 @@
 
 #include "data_generation/data_generator.h"
 #include "data_generation/prompt_formatting/prompt_formatter.h"
-#include "data_model/application_options.h"
+#include "data_model/models.h"
 #include "../services/prompting/prompt_directory.h"
 
 struct llama_model;
