--- tooling/pipeline/includes/data_model/generation_models.h
@@ -1,13 +1,0 @@
-#ifndef BIERGARTEN_PIPELINE_INCLUDES_DATA_MODEL_GENERATION_MODELS_H_
-#define BIERGARTEN_PIPELINE_INCLUDES_DATA_MODEL_GENERATION_MODELS_H_
-
-/**
- * @file data_model/generation_models.h
- * @brief Convenience include for shared generation payload models.
- */
-
-#include "data_model/brewery_location.h"
-#include "data_model/brewery_result.h"
-#include "data_model/user_result.h"
-
-#endif  // BIERGARTEN_PIPELINE_INCLUDES_DATA_MODEL_GENERATION_MODELS_H_
