--- tooling/pipeline/includes/data_model/pipeline_models.h
@@ -1,12 +1,0 @@
-#ifndef BIERGARTEN_PIPELINE_INCLUDES_DATA_MODEL_PIPELINE_MODELS_H_
-#define BIERGARTEN_PIPELINE_INCLUDES_DATA_MODEL_PIPELINE_MODELS_H_
-
-/**
- * @file data_model/pipeline_models.h
- * @brief Convenience include for pipeline-specific data models.
- */
-
-#include "data_model/enriched_city.h"
-#include "data_model/generated_brewery.h"
-
-#endif  // BIERGARTEN_PIPELINE_INCLUDES_DATA_MODEL_PIPELINE_MODELS_H_
