--- tooling/pipeline/includes/data_model/user_result.h
@@ -1,22 +1,0 @@
-#ifndef BIERGARTEN_PIPELINE_INCLUDES_DATA_MODEL_USER_RESULT_H_
-#define BIERGARTEN_PIPELINE_INCLUDES_DATA_MODEL_USER_RESULT_H_
-
-/**
- * @file data_model/user_result.h
- * @brief Generated user profile payload.
- */
-
-#include <string>
-
-/**
- * @brief Generated user profile payload.
- */
-struct UserResult {
-  /// @brief Username handle.
-  std::string username{};
-
-  /// @brief Short user biography.
-  std::string bio{};
-};
-
-#endif  // BIERGARTEN_PIPELINE_INCLUDES_DATA_MODEL_USER_RESULT_H_
