+++ tooling/pipeline/includes/services/database/sqlite_export_service.h
@@ -11,7 +11,7 @@
 #include <string>
 #include <unordered_map>
 
-#include "data_model/application_options.h"
+#include "data_model/models.h"
 #include "../datetime/date_time_provider.h"
 #include "export_service.h"
 #include "sqlite_export_service_helpers.h"
